fix(api): repairs test-modulix-content gh action
tests executed by the action need to run w/o database
