+++ api/tests/devcomp/integration/infrastructure/repositories/element-repository_test.js
@@ -1,12 +1,12 @@
 import sinon from 'sinon';
 
-import { QCUForAnswerVerification } from '../../../../src/devcomp/domain/models/element/QCU-for-answer-verification.js';
-import { Module } from '../../../../src/devcomp/domain/models/module/Module.js';
-import moduleDatasource from '../../../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
-import * as elementRepository from '../../../../src/devcomp/infrastructure/repositories/element-repository.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { expect } from '../../../test-helper.js';
-import { catchErr } from '../../../tooling/test-utils/error.js';
+import { QCUForAnswerVerification } from '../../../../../src/devcomp/domain/models/element/QCU-for-answer-verification.js';
+import { Module } from '../../../../../src/devcomp/domain/models/module/Module.js';
+import moduleDatasource from '../../../../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
+import * as elementRepository from '../../../../../src/devcomp/infrastructure/repositories/element-repository.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
+import { expect } from '../../../../test-helper.js';
+import { catchErr } from '../../../../tooling/test-utils/error.js';
 
 describe('Integration | DevComp | Repositories | ElementRepository', function () {
   describe('#getByIdForAnswerVerification', function () {
