+++ api/tests/devcomp/integration/infrastructure/repositories/module-issue-report-repository_test.js
@@ -1,7 +1,7 @@
-import { ModuleIssueReport } from '../../../../src/devcomp/domain/models/module/ModuleIssueReport.js';
-import * as moduleIssueReportRepository from '../../../../src/devcomp/infrastructure/repositories/module-issue-report-repository.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder, knex } from '../../../tooling/databases.js';
+import { ModuleIssueReport } from '../../../../../src/devcomp/domain/models/module/ModuleIssueReport.js';
+import * as moduleIssueReportRepository from '../../../../../src/devcomp/infrastructure/repositories/module-issue-report-repository.js';
+import { expect } from '../../../../test-helper.js';
+import { databaseBuilder, knex } from '../../../../tooling/databases.js';
 
 describe('Integration | DevComp | Repositories | ModuleIssueReportRepository', function () {
   describe('#save', function () {
