+++ api/tests/devcomp/integration/infrastructure/repositories/module-metadata-repository_test.js
@@ -1,13 +1,13 @@
 import sinon from 'sinon';
 
-import { ModuleDoesNotExistError } from '../../../../src/devcomp/domain/errors.js';
-import { Module } from '../../../../src/devcomp/domain/models/module/Module.js';
-import { ModuleMetadata } from '../../../../src/devcomp/domain/models/module/ModuleMetadata.js';
-import moduleDatasource from '../../../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
-import * as moduleMetadataRepository from '../../../../src/devcomp/infrastructure/repositories/module-metadata-repository.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { expect } from '../../../test-helper.js';
-import { catchErr } from '../../../tooling/test-utils/error.js';
+import { ModuleDoesNotExistError } from '../../../../../src/devcomp/domain/errors.js';
+import { Module } from '../../../../../src/devcomp/domain/models/module/Module.js';
+import { ModuleMetadata } from '../../../../../src/devcomp/domain/models/module/ModuleMetadata.js';
+import moduleDatasource from '../../../../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
+import * as moduleMetadataRepository from '../../../../../src/devcomp/infrastructure/repositories/module-metadata-repository.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
+import { expect } from '../../../../test-helper.js';
+import { catchErr } from '../../../../tooling/test-utils/error.js';
 
 describe('Integration | DevComp | Repositories | ModuleRepository', function () {
   describe('#getAllByIds', function () {
