+++ api/tests/devcomp/integration/infrastructure/repositories/passage-repository_test.js
@@ -1,11 +1,11 @@
 import sinon from 'sinon';
 
-import { Passage } from '../../../../src/devcomp/domain/models/Passage.js';
-import * as passageRepository from '../../../../src/devcomp/infrastructure/repositories/passage-repository.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder, knex } from '../../../tooling/databases.js';
-import { catchErr } from '../../../tooling/test-utils/error.js';
+import { Passage } from '../../../../../src/devcomp/domain/models/Passage.js';
+import * as passageRepository from '../../../../../src/devcomp/infrastructure/repositories/passage-repository.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
+import { expect } from '../../../../test-helper.js';
+import { databaseBuilder, knex } from '../../../../tooling/databases.js';
+import { catchErr } from '../../../../tooling/test-utils/error.js';
 
 describe('Integration | DevComp | Repositories | PassageRepository', function () {
   describe('#save', function () {
